CI/CD+Deps+libdnxpp: Use fmtlib/fmt for compatibility
Optionally depend on fmtlib/fmt to maintain compatibility with other
C++20 capable compilers that don't yet support the <format> header.

+ Added a new 'fmtlib' profile to vcpkg.json to conditionally grab
  fmtlib/fmt.
+ Added a new 'USE_FMTLIB' CMake option to conditionally look for and
  depend on fmtlib/fmt.
+ Added a new 'USE_FMTLIB' define to conditionally swap between using
  std::format and fmt::format.

NOTE: Currently this works by just aliasing fmt:: to std::, this is not
a great way of doing this, and ideally we should have an abstraction for
it in 'common.hpp', perhaps diannex::format?