libdnxpp+Sample: Add `registerFunctor` to address #6
This adds a new method to `DxInterpreter` called `registerFunctor`. It works
like the regular `registerFunction` however it's main difference is that
it's made with Functor Objects in mind, and thus, unlike it's function
variant, it won't make a copy of the object you pass in. Thus fixing the
issue described in #6.