CMake+Tests: Add Unit Testing for libdnxpp for #5
This adds Unit Testing with doctest as suggested by #5, it uses the
'standard' option for controlling builds `BUILD_TESTING`, which is also
behind a project check to make sure we aren't building tests when
libdnxpp is used as a subproject.

Right now the only tests are just a testified version of the Sample
project, more tests need to be made, but this is an alright start.